services/cv-pipeline/requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
redis>=5.0.0
psycopg2-binary>=2.9.0
boto3>=1.29.0
yt-dlp>=2023.12.0
